{{cookiecutter.repo}}/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
bcrypt==3.2.0
    # via -r requirements.in
cffi==1.14.4
    # via bcrypt
click==7.1.2
    # via uvicorn
fastapi==0.61.1
    # via -r requirements.in
gunicorn==20.0.4
    # via -r requirements.in
h11==0.11.0
    # via uvicorn
passlib==1.7.4
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydantic==1.6.1
    # via fastapi
pyjwt==1.7.1
    # via -r requirements.in
python-dotenv==0.14.0
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
six==1.15.0
    # via
    #   bcrypt
    #   python-multipart
starlette==0.13.6
    # via fastapi
uvicorn==0.12.2
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
